refactor: change the jwt logic to common module
