+++ package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@nestjs/common": "^10.3.8",
+    "@nestjs/config": "^3.2.2",
     "@nestjs/core": "^10.0.0",
     "@nestjs/jwt": "10.1.0",
     "@nestjs/mapped-types": "*",
