+++ src/documents/decorators/own-document.decorator.ts
@@ -1,9 +1,9 @@
 import { UseGuards, applyDecorators } from "@nestjs/common";
-import { AuthGuard } from "src/auth/guard/auth.guard";
-import { RolesGuard } from "src/auth/guard/roles.guard";
+import { AuthGuard } from "src/common/guard/auth.guard";
+import { RolesGuard } from "src/common/guard/roles.guard";
 import { Role } from "src/common/enums/rol.enum";
 import { OwnDocumentGuard } from "../guard/own-document.guard";
-import { Roles } from "src/auth/decorators/roles.decorators";
+import { Roles } from "src/common/decorators/roles.decorators";
 
 /**
  * Create a new decorator called AuthOwnDocument that takes an array of roles as an argument. This decorator will apply the Roles, UseGuards, and OwnDocumentGuard decorators to the route handler.
