+++ src/documents/documents.module.ts
@@ -5,12 +5,14 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Document } from './entities/document.entity';
 import { ProjectsModule } from '../projects/projects.module';
 import { UsersModule } from 'src/users/users.module';
+import { CommonModule } from 'src/common/common.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([Document]),
     ProjectsModule,
     UsersModule,
+    CommonModule,
   ],
   controllers: [DocumentsController],
   providers: [DocumentsService],
