+++ src/users/users.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/commo
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { Auth } from '../auth/decorators/auth.decorators';
+import { Auth } from '../common/decorators/auth.decorators';
 import { Role } from '../common/enums/rol.enum';
 
 @Controller('users')
